==> analysis/02_ML_benchmark/05_create_Llama3_matrix.R <==
# This code create the three dimensional array corresponding to Llama3+LCA, Llama3+SCA and Llama3+SLCA classifications from the generated txt files.


# Create the standard array skeleton that will be the storage template
matrix_result_template = array(data=0, dim=c(
  length(index),
  length(category),
  length(tone)
), 
dimnames = list(index,category,tone))


# -------------------------------------------------------------------------------------------------------------
# step 1 : Llama-3 alone
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for Llama3 standalone
Llama3 = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.


# Listing the paths
file_paths = data.frame(
  index = index,
  path = paste0("analysis/02_ML_benchmark/data/06_Llama3/llm_queries/output_1/initial_classification_result_",index,".json")
)

for(i in 1:length(index)){
  
  # Gather the LLM classification :
  suppressWarnings({
    result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
  })
  result = jsonlite::fromJSON(result)
  
  # Extract the output from the result
  result = result$output
  
  # erase the themes, keep the categories
  result_2 = list() # json data, but without themes
  themes = ls(result)
  for(current_theme in themes){
    result_2 = append(result_2, result[[current_theme]])
  }
  names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
  result = result_2 # Save the trimmed json in result
  
  # Store the prediction in the 3D matrix
  for(current_category in category){
    Llama3[i,current_category,"positive"] = ifelse(result[[current_category]][["positive"]]!="",1,0)
    Llama3[i,current_category,"negative"] = ifelse(result[[current_category]][["negative"]]!="",1,0)
  }
}


saveRDS(Llama3,"analysis/02_ML_benchmark/data/06_Llama3/Llama3.rds")



# -------------------------------------------------------------------------------------------------------------
# step 2 : Llama-3 + SCA
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for Llama3_SCA
Llama3_SCA = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.

# Listing the paths
file_paths = data.frame(
  index = index,
  path = paste0("analysis/02_ML_benchmark/data/06_Llama3/evaluations/sca/result_",index,".json")
)

for(i in 1:length(index)){
  
  # Gather the LLM classification :
  suppressWarnings({
    result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
  })
  result = jsonlite::fromJSON(result)
  
  # Extract the output from the result
  result = result$output
  
  # erase the themes, keep the categories
  result_2 = list() # json data, but without themes
  themes = ls(result)
  for(current_theme in themes){
    result_2 = append(result_2, result[[current_theme]])
  }
  names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
  result = result_2 # Save the trimmed json in result
  
  # Store the prediction in the 3D matrix
  for(current_category in category){
    Llama3_SCA[i,current_category,"positive"] = result[[current_category]][["positive"]]
    Llama3_SCA[i,current_category,"negative"] = result[[current_category]][["negative"]]
  }
}


saveRDS(Llama3_SCA,"analysis/02_ML_benchmark/data/06_Llama3/Llama3_SCA.rds")



# -------------------------------------------------------------------------------------------------------------
# step 3 : Llama-3 + LCA
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for Llama3+LCA
Llama3_LCA = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.


# Listing the paths
file_paths = data.frame(
  index = index,
  path = paste0("analysis/02_ML_benchmark/data/06_Llama3/evaluations/lca/output_1/result_",index,".json")
)

for(i in 1:length(index)){
  
  # Gather the LLM classification :
  suppressWarnings({
    result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
  })
  result = jsonlite::fromJSON(result)
  
  # Extract the output from the result
  result = result$output
  
  # erase the themes, keep the categories
  result_2 = list() # json data, but without themes
  themes = ls(result)
  for(current_theme in themes){
    result_2 = append(result_2, result[[current_theme]])
  }
  names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
  result = result_2 # Save the trimmed json in result
  
  # Store the prediction in the 3D matrix
  for(current_category in category){
    Llama3_LCA[i,current_category,"positive"] = result[[current_category]][["positive"]]
    Llama3_LCA[i,current_category,"negative"] = result[[current_category]][["negative"]]
  }
}


saveRDS(Llama3_LCA,"analysis/02_ML_benchmark/data/06_Llama3/Llama3_LCA.rds")



# -------------------------------------------------------------------------------------------------------------
# step 4 : Llama-3 + SLCA
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for Llama3+SLCA
Llama3_SLCA = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.


# Listing the paths
file_paths = data.frame(
  index = index,
  path = paste0("analysis/02_ML_benchmark/data/06_Llama3/evaluations/slca/result_",index,".json")
)

for(i in 1:length(index)){
  
  # Gather the LLM classification :
  suppressWarnings({
    result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
  })
  result = jsonlite::fromJSON(result)
  
  # Extract the output from the result
  result = result$output
  
  # erase the themes, keep the categories
  result_2 = list() # json data, but without themes
  themes = ls(result)
  for(current_theme in themes){
    result_2 = append(result_2, result[[current_theme]])
  }
  names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
  result = result_2 # Save the trimmed json in result
  
  # Store the prediction in the 3D matrix
  for(current_category in category){
    Llama3_SLCA[i,current_category,"positive"] = result[[current_category]][["positive"]]
    Llama3_SLCA[i,current_category,"negative"] = result[[current_category]][["negative"]]
  }
}


saveRDS(Llama3_SLCA,"analysis/02_ML_benchmark/data/06_Llama3/Llama3_SLCA.rds")
